requirements.txt
apache-airflow
apache-airflow-providers-snowflake
snowflake-connector-python
pandas
requests